refactor: move validateContract server test helper to testHelpers.ts
