+++ backend/tests/rooms.test.ts
@@ -1,14 +1,10 @@
-import assert from "assert";
 import supertest from "supertest";
 import app from "../src/app";
 import { contracts } from "../src/data/contracts";
-import { departments } from "../src/data/departments";
-import { people } from "../src/data/people";
 import { rooms } from "../src/data/rooms";
-import { titles } from "../src/data/titles";
 import type { Contract } from "../src/models";
 import { createAllTables, dropAllTables, seedData } from "../src/seed";
-import type { ExpectedContract } from "../src/types/other";
+import { validateContract } from "./testHelpers";
 
 const api = supertest(app);
 
@@ -17,36 +13,6 @@ beforeEach(async () => {
   await createAllTables();
   await seedData();
 });
-
-const validateContract = (
-  contract: Contract,
-  expectedContract: ExpectedContract,
-) => {
-  const expectedPerson = people.find(
-    (person) => person.id === expectedContract.personId,
-  );
-  const expectedDepartment = departments.find(
-    (dept) => dept.id === expectedPerson?.departmentId,
-  );
-  const expectedTitle = titles.find(
-    (title) => title.id === expectedPerson?.titleId,
-  );
-
-  assert(expectedPerson, "expected person should be defined");
-  assert(expectedDepartment, "expected department should be defined");
-  assert(expectedTitle, "expected title should be defined");
-
-  expect(contract).toMatchObject({
-    startDate: expectedContract.startDate.toISOString().slice(0, 10),
-    endDate: expectedContract.endDate.toISOString().slice(0, 10),
-    person: {
-      firstName: expectedPerson.firstName,
-      lastName: expectedPerson.lastName,
-      department: { name: expectedDepartment.name },
-      title: { name: expectedTitle.name },
-    },
-  });
-};
 
 test("rooms data is returned as json", async () => {
   await api
